Move enable-editor-hotkeys hook to extra file
